// src/main/webui/src/app/auth/enums/o-auth-event-type.ts
export enum OAuthEventType {
  DISCOVERY_DOCUMENT_LOADED = 'discovery_document_loaded',
  JWKS_LOAD_ERROR = 'jwks_load_error',
  INVALID_NONCE_IN_STATE = 'invalid_nonce_in_state',
  DISCOVERY_DOCUMENT_LOAD_ERROR = 'discovery_document_load_error',
  DISCOVERY_DOCUMENT_VALIDATION_ERROR = 'discovery_document_validation_error',
  USER_PROFILE_LOADED = 'user_profile_loaded',
  USER_PROFILE_LOAD_ERROR = 'user_profile_load_error',
  TOKEN_RECEIVED = 'token_received',
  TOKEN_ERROR = 'token_error',
  CODE_ERROR = 'code_error',
  TOKEN_REFRESHED = 'token_refreshed',
  TOKEN_REFRESH_ERROR = 'token_refresh_error',
  SILENT_REFRESH_ERROR = 'silent_refresh_error',
  SILENTLY_REFRESHED = 'silently_refreshed',
  SILENT_REFRESH_TIMEOUT = 'silent_refresh_timeout',
  TOKEN_VALIDATION_ERROR = 'token_validation_error',
  TOKEN_EXPIRES = 'token_expires',
  SESSION_CHANGED = 'session_changed',
  SESSION_ERROR = 'session_error',
  SESSION_TERMINATED = 'session_terminated',
  SESSION_UNCHANGED = 'session_unchanged',
  LOGOUT = 'logout',
  POPUP_CLOSED = 'popup_closed',
  POPUP_BLOCKED = 'popup_blocked',
  TOKEN_REVOKE_ERROR = 'token_revoke_error',
}